Move files to Auth.Common namespace.
