+++ src/GovITHub.Auth.Common/Data/ApplicationDbContext.cs
@@ -1,14 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
-using GovITHub.Auth.Identity.Models;
-using GovITHub.Auth.Identity.Services.Audit.DataContracts;
-using GovITHub.Auth.Identity.Data.Models;
-using GovITHub.Auth.Identity.Services.DeviceDetection.DataContracts;
-//using GovITHub.Auth.Identity.Models;
+using GovITHub.Auth.Common.Services.DeviceDetection.DataContracts;
 using GovITHub.Auth.Common.Models;
 using GovITHub.Auth.Common.Services.Audit.DataContracts;
 using GovITHub.Auth.Common.Data.Models;
