+++ src/GovITHub.Auth.Common/Services/DeviceDetection/DeviceInfoBuilders/Extensions/MatchExtensions.cs
