+++ src/GovITHub.Auth.Common/Services/DeviceDetection/DeviceInfoBuilders/YamlSchema/DeviceModel.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace GovITHub.Auth.Common.Services.DeviceDetection.DeviceInfoBuilders.YamlSchema
+{
+    public class DeviceModel
+    {
+        public string Regex { get; set; }
+
+        public string Model { get; set; }
+
+        public string Device { get; set; }
+    }
+}
