+++ src/GovITHub.Auth.Common/Services/DeviceDetection/DeviceInfoBuilders/YamlSchema/Engine.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace GovITHub.Auth.Common.Services.DeviceDetection.DeviceInfoBuilders.YamlSchema
+{
+    [DebuggerDisplay("{Default}")]
+    public class Engine
+    {
+        public string Default { get; set; }
+    }
+}
