+++ src/GovITHub.Auth.Common/Services/DeviceDetection/ILoginDeviceManagementService.cs
@@ -1,0 +1,10 @@
+﻿using GovITHub.Auth.Common.Services.DeviceDetection.DataContracts;
+using System.Threading.Tasks;
+
+namespace GovITHub.Auth.Common.Services.DeviceDetection
+{
+    public interface ILoginDeviceManagementService
+    {
+        Task RegisterDeviceLoginAsync(string userId, DeviceInfo deviceInfo);
+    }
+}
