+++ src/GovITHub.Auth.Common/project.json
@@ -25,17 +25,30 @@
     "MailKit": "1.10.0",
     "WebApiContrib.Core.Formatter.Csv": "1.0.0",
     "SapientGuardian.MySql.Data": "6.9.813",
-    "SapientGuardian.EntityFrameworkCore.MySql": "7.1.14"
+    "SapientGuardian.EntityFrameworkCore.MySql": "7.1.14",
+    "YamlDotNet": "4.0.0"
+  },
+
+  "buildOptions": {
+    "emitEntryPoint": true,
+    "preserveCompilationContext": true,
+    "embed": {
+      "includeFiles": [
+        "../../dependencies/device-detector/regexes/client/browsers.yml",
+        "../../dependencies/device-detector/regexes/device/mobiles.yml",
+        "../../dependencies/device-detector/regexes/oss.yml"
+      ]
+    }
   },
 
   "frameworks": {
-        "netcoreapp1.0": {
-          "imports": [
-            "dotnet5.6",
-            "dnxcore50",
-            "net46",
-            "portable-net45+win8"
-          ]
-        }
+    "netcoreapp1.0": {
+      "imports": [
+        "dotnet5.6",
+        "dnxcore50",
+        "net46",
+        "portable-net45+win8"
+      ]
+    }
   }
 }
