--- src/GovITHub.Auth.Identity/Services/DeviceDetection/DeviceInfoBuilders/YamlSchema/BrowserRegex.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace GovITHub.Auth.Identity.Services.DeviceDetection.DeviceInfoBuilders.YamlSchema
-{
-    [DebuggerDisplay("{Name}")]
-    public class BrowserRegex : IRegex
-    {
-        public string Regex { get; set; }
-        public string Name { get; set; }
-        public string Version { get; set; }
-        public Engine Engine { get; set; }
-    }
-}
