--- src/GovITHub.Auth.Identity/Services/DeviceDetection/DeviceInfoBuilders/YamlSchema/Engine.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace GovITHub.Auth.Identity.Services.DeviceDetection.DeviceInfoBuilders.YamlSchema
-{
-    [DebuggerDisplay("{Default}")]
-    public class Engine
-    {
-        public string Default { get; set; }
-    }
-}
