+++ test/GovITHub.Auth.Identity.Tests/project.json
@@ -2,13 +2,10 @@
   "version": "1.0.0-*",
   "testRunner": "xunit",
   "dependencies": {
-    "Microsoft.NETCore.App": {
-      "version": "1.0.1",
-      "type": "platform"
-    },
+    "Microsoft.NETCore.App": "1.1.0",
     "xunit": "2.1.0",
     "dotnet-test-xunit": "2.2.0-preview2-build1029",
-    "GovITHub.Auth.Identity": "1.0.0-*"
+    "GovITHub.Auth.Common": "1.0.0-*"
   },
 
   "frameworks": {
